ci: add static analysis, script linting, sanitizers, -Werror (Tier A)
Harden CI and the build without changing firmware behavior:

- Warnings-as-errors, opt-in via -DDUTLER_WERROR (CI on, local off). Scope the
  strict set (-Wshadow -Wdouble-promotion -Wconversion on top of -Wall -Wextra)
  to our OWN sources via set_source_files_properties — a target-level option
  would leak onto the Pico SDK / TinyUSB sources (attached through INTERFACE
  libs), which aren't clean under it.
- Host unit tests now build -Werror and run under AddressSanitizer + UBSan
  (-DDUTLER_SANITIZE).
- New CI jobs: shellcheck + ruff (script linting) and cppcheck (static
  analysis), both pinned/conservative to stay reproducible.
- Build job: -Werror and an arm-none-eabi-size footprint report.
- Workflow hygiene: least-privilege `permissions: contents: read` and
  `concurrency` cancel-in-progress.

Fixes surfaced by the new checks:
- command.c: print "%u" with (unsigned)(i + 1) (cppcheck invalidPrintfArgType).
- env.sh: `# shellcheck shell=bash`; split declare/assign (SC2148/SC2155).
- flash.sh: silence SC1091 on the runtime `source ./env.sh`.
- tools/*.py: split one-line imports and try/except blocks (ruff E401/E70x).

CONTRIBUTING documents the checks and how to run them locally. Verified:
all linters clean, host tests pass under sanitizers, firmware builds -Werror.

+++ src/command.c
@@ -29,9 +29,9 @@ static void print_status(void) {
         const char *nm = g_settings.out_name[i];
         const char *st = outputs_get(i) ? "on" : "off";
         if (nm[0])
-            snprintf(msg, sizeof(msg), "out %u (%s) %s\r\n", i + 1, nm, st);
+            snprintf(msg, sizeof(msg), "out %u (%s) %s\r\n", (unsigned)(i + 1), nm, st);
         else
-            snprintf(msg, sizeof(msg), "out %u %s\r\n", i + 1, st);
+            snprintf(msg, sizeof(msg), "out %u %s\r\n", (unsigned)(i + 1), st);
         console_print(msg);
     }
     char pc = g_settings.parity == 1 ? 'O' : g_settings.parity == 2 ? 'E' : 'N';
